Split into multiple *.Types.hs modules
